SuperAdmin_ProfSpace_AdminSpace

+++ etudiant_controller.php
@@ -19,12 +19,63 @@ if (isset($_POST['modiferEtudiant'])) {
     $annee_diplome = $_POST['annee_diplome'];
     $specialite = $_POST['specialite'];
     $parcours_souhaite = $_POST['parcours_souhaite'];
-    $filename = $_FILES['image']['name'];
-    $tmp_name = $_FILES['image']['tmp_name'];
-    $destination = 'upload/images/' . $filename;
-    move_uploaded_file($tmp_name, $destination);
+    $mois_stage = $_POST['mois_stage'];
+    $experience = $_POST['experience'];
+    $mot_de_passe = $_POST['ancien_mot_de_passe'];
+    $nouveau_mot_de_passe = $_POST['nouveau_mot_de_passe'];
 
-    if (!empty($etudiant_nom) && !empty($etudiant_prenom) && !empty($cin) && !empty($email) && !empty($tele) && !empty($adresse)) {
+    if (!empty($mot_de_passe) && !empty($nouveau_mot_de_passe)) {
+        $query = "SELECT * FROM etudiants WHERE EtudiantID = $etudiant_id AND Password = '$mot_de_passe'";
+        $query_run = mysqli_query($connection, $query);
+
+        if (mysqli_num_rows($query_run) > 0) {
+            $query = "UPDATE etudiants SET Password = '$nouveau_mot_de_passe' WHERE EtudiantID = $etudiant_id";
+            $query_run = mysqli_query($connection, $query);
+        } else {
+            $_SESSION['status'] = "Mot de passe incorrect";
+            header('location: etudiant_espace.php');
+            exit;
+        }
+    }
+
+    $filename = $_FILES['image']['name'];
+
+    if (!empty($filename)) {
+        $tmp_name = $_FILES['image']['tmp_name'];
+        $destination = 'upload/images/' . $filename;
+        $imageFileType = strtolower(pathinfo($destination, PATHINFO_EXTENSION));
+        if ($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg") {
+            $_SESSION['status'] = "Le fichier doit être une image";
+            header('location: etudiant_espace.php');
+            exit;
+        } else if ($_FILES['image']['size'] > 6 * 1024 * 1024) {
+            $_SESSION['status'] = "Le fichier doit être moins de 6 Mo";
+            header('location: etudiant_espace.php');
+            exit;
+        } else {
+            // Vérification si le fichier existe déjà
+            if (!file_exists($destination)) {
+                // Déplacer le fichier vers sa destination avec le nouveau nom
+                move_uploaded_file($tmp_name, $destination);
+                $query = "UPDATE etudiants SET Image = '$filename' WHERE EtudiantID = $etudiant_id";
+                $query_run = mysqli_query($connection, $query);
+            } else {
+                // Renommer le fichier avec un nom unique
+                $new_filename = uniqid() . '_' . $filename;
+                $destination = 'upload/images/' . $new_filename;
+                move_uploaded_file($tmp_name, $destination);
+                $query = "UPDATE etudiants SET Image = '$new_filename' WHERE EtudiantID = $etudiant_id";
+                $query_run = mysqli_query($connection, $query);
+            }
+        }
+    }
+
+    //print all inputs
+    // foreach ($_POST as $key => $value) {
+    //     echo $key . " : " . $value . "<br>";
+    // }
+
+    if (!empty($etudiant_nom) && !empty($etudiant_prenom) && !empty($cin) && !empty($tele)){
 
         $query = "UPDATE etudiants SET 
             Etudiant_name = '$etudiant_nom',
